Disable DYNAMIC_FILTER feature if gyro loop is less than 2KHz The dynamic filter is not usefull with gyro loop times less than 2KHz so disable the feature and prevent the code from running in this circumstance.
In `isDynamicFilterActive()` we can't just rely on the feature because the feature will be disabled after the gyro has initialized and is running - possibly running into the underflow/wedging problem. And unfortunately we can't disable the feature earlier because we need to know the `targetLooptime` which is determined during gyro init.